pyproject.toml
[tool.poetry]
name = "blast-courses"
version = "0.1.0"
description = "The platform will collaborate with companies that offer ready-made courses. Courses from various fields of military science will be available, including tactics, security, battlefield medicine, communications, etc. Currently, we are focusing on courses in tactical medicine and courses for civilian adaptation."
authors = ["marik <letschuka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
gunicorn = "^22.0.0"
djoser = "^2.2.2"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
black = "^24.4.2"
google-auth = "^2.29.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    | migrations
    | .git
    | .venv
    | env
    | build
    | dist
)/
'''
